+++ AdlsDotNetSDKUnitTest/SdkUnitTest.cs
@@ -146,7 +146,7 @@ namespace Microsoft.Azure.DataLake.Store.UnitTest
                 new AclEntry(AclType.user, _nonOwner3ObjectId, AclScope.Access, AclAction.ExecuteOnly),
                 new AclEntry(AclType.group, Group1Id, AclScope.Access, AclAction.ExecuteOnly)
             };
-            _adlsClient.ModifyAclEntries("/",nonOwnerAclSpec);
+            _adlsClient.ModifyAclEntries("/", nonOwnerAclSpec);
             _adlsClient.ModifyAclEntries(UnitTestDir, nonOwnerAclSpec);
         }
         #region Setup
@@ -1029,7 +1029,7 @@ namespace Microsoft.Azure.DataLake.Store.UnitTest
         /// </summary>
         /// <param name="overwrite">Whether to overwrite the existing destination if it exists</param>
         [TestMethod]
-        [DataRow(true)]
+        [DataRow(false)]
         [DataRow(true)]
         public void TestRenameDirectoryDestinationExist(bool overwrite)
         {
@@ -1375,6 +1375,8 @@ namespace Microsoft.Azure.DataLake.Store.UnitTest
         {
             TestConcatTwoFile(false, UnitTestDir + "/destPath2.txt", UnitTestDir + "/Source");
             TestConcatTwoFile(true, UnitTestDir + "/destPath3.txt", UnitTestDir + "/Source1");
+            TestConcatTwoFile(false, UnitTestDir + "/destPath6.txt", UnitTestDir + "/Source1", "prefix+with,signs");
+            TestConcatTwoFile(true, UnitTestDir + "/destPath7.txt", UnitTestDir + "/Source1", "prefix+with,signs");
         }
 
         /// <summary>
@@ -1383,17 +1385,17 @@ namespace Microsoft.Azure.DataLake.Store.UnitTest
         /// <param name="deleteSource">Whether to delete source directory</param>
         /// <param name="destPath">Destination filename</param>
         /// <param name="sourcePath">Source directory</param>
-        public void TestConcatTwoFile(bool deleteSource, string destPath, string sourcePath)
+        public void TestConcatTwoFile(bool deleteSource, string destPath, string sourcePath, string sourceFileNamePrefix = "")
         {
             List<string> srcList = new List<string>();
-            string srcFile1 = sourcePath + "/srcPath1.txt";
+            string srcFile1 = sourcePath + "/" + sourceFileNamePrefix + "srcPath1.txt";
             string text1 = RandomString(2 * 1024 * 1024);
             byte[] textByte1 = Encoding.UTF8.GetBytes(text1);
             using (var ostream = _adlsClient.CreateFile(srcFile1, IfExists.Overwrite, ""))
             {
                 ostream.Write(textByte1, 0, textByte1.Length);
             }
-            string srcFile2 = sourcePath + "/srcPath2.txt";
+            string srcFile2 = sourcePath + "/" + sourceFileNamePrefix + "srcPath2.txt";
             string text2 = RandomString(3 * 1024 * 1024);
             byte[] textByte2 = Encoding.UTF8.GetBytes(text2);
             using (var ostream = _adlsClient.CreateFile(srcFile2, IfExists.Overwrite, ""))